% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compeval.R
\name{compeval}
\alias{compeval}
\title{Provides a range of statistics for evaluating how components are related to a SOM model.}
\usage{
compeval(som_obj, labsize = 1, colormod = NULL)
}
\arguments{
\item{som_obj}{an object returned from sommix_summ}

\item{labsize}{sets label size on plots}

\item{colormod}{sets colors for plots}
}
\value{
a list containing component evaluation statistics
\itemize{
 \item {Component_Fit} a dataframe of model fit statistics for each component/variable
 \item {Component_Map} a dataframe of Moran's I statistics for each component/variable
}
}
\description{
Calculates a range of statistics for components (i.e., variables) of the training data in order to better understand relationships with the SOM model.
}
\details{
Understanding how well components relate to the SOM model can provide insights into variable importance. Here correlations, root means square error, and mean absolute error are calculated using observed vs predicted values from the SOM model. Measures of spatial autocorrelation are provided to enhance understanding of spatial patterning among components on the SOM.
}
