% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coreval.R
\name{coreval}
\alias{coreval}
\title{A tool to evaluate and plot correlation structure}
\usage{
coreval(x, cormeth = "spearman", labsize = 1, varlabs = NULL)
}
\arguments{
\item{x}{is a dataframe object}

\item{cormeth}{passes correlation method to cor()}

\item{labsize}{sets label size}

\item{varlabs}{specifies variable labels}
}
\value{
a correlation matrix
}
\description{
coreval explores data correlations
}
\examples{
#NIEHS Mixtures Workshop dataset1
data(dataset1)
coreval(dataset1[,2:9])
}
