% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset2.R
\docType{data}
\name{dataset2}
\alias{dataset2}
\title{Data Set 2 Mixture Simulated Data with an Environmentally Relevant Complex Correlation Pattern}
\format{
A data frame with 500 rows and 18 variables:
\describe{
  \item{OBS}{subject identifier}
  \item{Y}{The outcome, a continuous variable}
  \item{X1}{a continous exposure biomarker}
  \item{X2}{a continous exposure biomarker}
  \item{X3}{a continous exposure biomarker}
  \item{X4}{a continous exposure biomarker}
  \item{X5}{a continous exposure biomarker}
  \item{X6}{a continous exposure biomarker}
  \item{X7}{a continous exposure biomarker}
  \item{X8}{a continous exposure biomarker}
  \item{X9}{a continous exposure biomarker}
  \item{X10}{a continous exposure biomarker}
  \item{X11}{a continous exposure biomarker}
  \item{X12}{a continous exposure biomarker}
  \item{X13}{a continous exposure biomarker}
  \item{X14}{a continous exposure biomarker}
  \item{Z1}{a continous confounder}
  \item{Z2}{a continous confounder}
  \item{Z3}{a binary confounder}
  ...
}
}
\source{
\url{https://www.niehs.nih.gov/about/events/pastmtg/2015/statistical/index.cfm}
}
\usage{
dataset2
}
\description{
This data file is intended to represent data from a cross-sectional study of 14 biomarkers (e.g., PCBs, dioxins, furans) from biomonitoring data potentially associated with a biomarker of effect (e.g., ALT as a marker of liver toxicity).
Three covariates are also included, one binary and two continuous.
}
\keyword{datasets}
