% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapsize.R
\name{mapsize}
\alias{mapsize}
\title{Determining SOM size}
\usage{
mapsize(
  x,
  kmn = NULL,
  kmx = NULL,
  itermax. = NULL,
  nstarts. = NULL,
  maptopo. = NULL,
  distmet. = NULL,
  lmode. = NULL,
  symsize = 1
)
}
\arguments{
\item{x}{is a data object to train the map. Should be a numerical or factor based data matrix.}

\item{kmn}{is the minimum number of map units (aka nodes) to evaluate. Default value is 2.}

\item{kmx}{is maximum number of map units (aka nodes) to evaluate. Default is 5*sqrt(n).}

\item{itermax.}{maximum number of iterations passed to sommix. Default is 500*number of map nodes (k).}

\item{nstarts.}{number of random initializatons passed to sommix. Default is 5.}

\item{maptopo.}{map topology passed to sommix. Default is rectangular.}

\item{distmet.}{distance method passed to sommix. Default is Euclidean.}

\item{lmode.}{m initializatons passed to sommix}

\item{symsize}{sets symbol size on plots}
}
\value{
Panels a-d on the diagnostic plot illustrate common model perfomance metrics as a function of map size. Panels e-f examine within-class-sum-of-squares and frequency distributions as a function of map size. A list of model and class-level perfomance statistics is also returned.
 \itemize{
 \item {SOMX} x dimension
 \item {SOMY} y dimension
 \item {K} Number of map nodes
 \item {R2} R2
 \item {ADJ_R2} Adjusted R2
 \item {MAE} mean absolute error based on class assignment distances
 \item {RMSE} root-mean-square-error based on class assignment distances
 \item {AIC} a form of Akaikes Information Criteria applied to clustering algorithms
 \item {TotWCSS}  Total Within-Cluster Sum-of-Squares
 \item {N} Number class assignments
 \item {FREQ} Proportion of class assignments
 \item {WCD} average within-class distance
 \item {BCD}  average between-class distacne
 \item {WB_Ratio}  WCD/BCD

}
}
\description{
mapsize provides diagnostic plots and summaries of select criteria for determining how the size of the SOM influences characteristics of the SOM model.
}
\details{
An important step in the application of SOM are the user provided inputs for the dimensions of the mapping (i.e., size). Here we provide common model performance metrics and class-level evaluations in effort to assist the user in determining an appropriate map size.
}
\examples{
#NIEHS Mixtures Workshop dataset1
data(dataset1)
mapsize(scale(dataset1[,3:9]), kmx=10, itermax.=10)
}
