% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcaeval.R
\name{pcaeval}
\alias{pcaeval}
\title{Principal Component Analysis with Diagnostic Plots}
\usage{
pcaeval(x, labsize = 1, colormod = NULL, varlabs = NULL, ...)
}
\arguments{
\item{x}{a data}

\item{labsize}{sets label size}

\item{colormod}{can be used to customize plot color schemes. The default uses terrain palletes from grDevices}

\item{varlabs}{specifies x-axis variable labels}

\item{...}{arguments passed to or from other methods}
}
\value{
a list of pca results with class "prcomp"
}
\description{
Performs principal component analysis on the data and returns the results as an object of class prcomp. Scree, variance, and loading plots are provided.
}
\details{
Assessing patterns of intercorrelations between x-variables is an important aspect in multivariate analysis. This assessment tool can assist with understanding the complexitiy of the underlying data by identifiying the important primary modes of variance that exist. In most The evaluation tool is based upon application of prcomp for stats. The
}
\examples{
#NIEHS Mixtures Workshop dataset1
data(dataset1)
pcaeval(scale(dataset1[,2:9]))
}
