% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somfitstats.R
\name{somFitStats}
\alias{somFitStats}
\title{Standard Model Fit Statistics applied to Self-Organizing Maps (SOM) model}
\usage{
somFitStats(som_obj)
}
\arguments{
\item{som_obj}{an object with class 'sommix'}
}
\value{
a dataframe of values

TSS     The total sum of squares

WCSS    The within-class sum of squares

TOTWCSS The total within-class sum of squares

BCSS    The between-class sum of squares, i.e. TSS-TOTWCSS

R2     The total variance explained by the SOM classes (BCSS/TSS)

ADJ_R2 The total variance explained by the SOM classes adjusted for class number

MAE     The mean absolute error (MAE) derived from the mean class assignment distance, i.e., quantization error

RMSE    The root-mean-square-error derived from class assignment distances, i.e., quantization error
}
\description{
somFitStats calculates common measures of model fit for a SOM model
}
