% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sommix.R
\name{sommix}
\alias{sommix}
\title{Kohonen's Self-Organizing Maps (SOM) algrithm}
\usage{
sommix(
  X,
  somx = 3,
  somy = 2,
  maptopo = NULL,
  itermax = NULL,
  seedopt = "Y",
  distmet = NULL,
  nstarts = NULL,
  lmode = NULL
)
}
\arguments{
\item{X}{is a data object to train the map. Should be a numerical or factor based data matrix. data.frame objects are not supported.}

\item{somx}{is x-dimension of the SOM}

\item{somy}{is y-dimension of the SOM}

\item{maptopo}{specifies the topology of the SOM grid as either "rectangular" or "hexagonal"}

\item{itermax}{specifies the number if learning iterations}

\item{seedopt}{specifies if optimal initialization values are to be used. "Y" or "N" accepted.}

\item{distmet}{specifies dissimilarity metric. Current options include "sumofsquares", "euclidean", "manhattan", and "tanimoto". Default is to use "Euclidean" for continuous data, and "tanimoto" for factors.}

\item{nstarts}{specifies number of initialization schemes to test if inits are not provided}

\item{lmode}{specifies the learning algorithm. The default is "online" but "batch" and "pbatch" are available via kohonen}
}
\value{
a 'som' object with class Kohonen
}
\description{
Kohonen's Self-Organizing Map algorithm is an artificial neural network (ANN) that applied competitive learning with an integraged neighborhood function in order to discover and visualize patterns in multivariate datasets. The 'map' is a low-dimensional representation that illustrates the discovered patterns (i.e., profiles) in a compact, spatially organized way. This unique feature allows for larger numbers of classes to be more easily understand, offer users the ability to construct high resolution classifications. This is a key distinction from traditional clustering/dimension reduction approaches which often seek to minimal representation.
}
\details{
Applies Kohonens self-organizing map via kohonen using robust tuning parameters and optimized initial values
}
\examples{
#NIEHS Mixtures Workshop dataset1
data(dataset1)
#Apply SOM
sommod<-sommix(scale(dataset1[,3:9]), somx=3, somy=2)
#Summarize Output
somsumm<-sommix_summ(sommod)
#Plot Results
sommix_bar(som_obj=sommod, varnames=NULL, colormod=NULL,nodelab=TRUE, labtype="IDS",
labsize=1, addFreq=TRUE, freqtype="frq", freqsize=1, legsymsize=2, leglabsize=1,
legtxtlas=2, barstat="MED")
#Evaluate SOM
sommix_eval(sommod, labtype = "IDs")
}
\references{
Kohonen, T. (1995) Self-Organizing Maps. Springer-Verlag
}
