% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sommixplot.R
\name{sommix_bar}
\alias{sommix_bar}
\title{Plots SOM profiles using bar charts}
\usage{
sommix_bar(
  som_obj,
  varnames = NULL,
  colormod = NULL,
  nodelab = TRUE,
  labtype = "IDS",
  labsize = 1,
  addFreq = FALSE,
  freqtype = "frq",
  freqsize = 1,
  legsymsize = 2,
  leglabsize = 1,
  legtxtlas = 2,
  barstat = "MED"
)
}
\arguments{
\item{som_obj}{a SOM object}

\item{varnames}{specifies variable names}

\item{colormod}{can be used to customize plot color schemes. The default is terrain.colors().}

\item{nodelab}{whether or not to display reference labels aon each node}

\item{labtype}{determines type of node referencing to display. "IDs" and "XYs" are available}

\item{labsize}{sets size of node labels}

\item{addFreq}{whether or not to display frequency labels}

\item{freqtype}{determines type of frequency labels. Relative frequencies can be specified with "frq" or counts with "cnts"}

\item{freqsize}{frequency label cex value}

\item{legsymsize}{legend symbol cex value}

\item{leglabsize}{legend text cex value}

\item{legtxtlas}{legend text position rotation passed to las. Numeric in {0,1,2,3}; the style of axis label.}

\item{barstat}{specifies summary statistic to plot as bars. "MED" and "AVG" are available}
}
\value{
a kohonen plot that uses barplots to describe the profiles for each map node.
}
\description{
Barplots are used to represent standardized summary values of components under each SOM profile using bars of various lengths.
}
\details{
Barplots are a straighforward way to show patterns in component values under each SOM profile. The size of the bar represents the component summary under each profile. Error bars are also provided.  between obser
}
