% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sommixplot.R
\name{sommix_eval}
\alias{sommix_eval}
\title{Generate multiple plots of SOM model results}
\usage{
sommix_eval(som_obj, labtype = "IDs", labsize = 1, legcex = 1, colormod = NULL)
}
\arguments{
\item{som_obj}{a SOM object}

\item{labtype}{determines label type. "IDs" and "XYs" available.}

\item{labsize}{sets label size}

\item{legcex}{sets legend cex}

\item{colormod}{sets colorscheme}
}
\value{
Panel a-c plot SOM profile results using a range of techniques designed to enhance understanding of results. A list is also returned that includes the following items.
 \itemize{
 \item {EVALTAB} a dataframe containing classification assignments for SOM, kmeans (KM), Ward (HC). Principal component scores for PC1 and PC2 are also included.
 \item {SOM_Profiles} SOM profiles (a.k.a., codebook vectors)
 \item {KM_Profiles} K-means profiles (a.k.a., cluster centers)
 \item {HC_Profiles} Ward's hierarchical clustering profiles (a.k.a., cluster centers)
 \item {PC1_Loadings} loadings from the first principal component
 \item {PC2_Loadings} loadings from the second principal component
 \item {CLASSIC_AGREEMENT} measures of class agreement ranging from 0:1 where 1 reflects perfect agreement
}
}
\description{
Plots SOM output using Parallel Coordinate Plots, Multi-dimensional Scaling (MDS), and Sammons mapping.  kmeans, wards, and PCA
}
\details{
Additional visualization of SOM results can lead to stronger interpretion. Panel (a) presenst a parallel coordinates plot of SOM profile results in order visualize differences in the individual components. Each vertical bar represents a variable axis where minimum values are on the bottom and maximum values are on top. A line is then used to connect component values for each profile. Panel (b) presents a multidimensional scaling plot that allows visualization of how well the SOM profile results span the original data space and how they compare to profiles generated from kmeans and Ward's heirarchical clustering. Panel (c) presents a Sammon's mapping of SOM profiles that effectively visualizes the the SOM using a distance preserving map projection. This provides a clearer picture of similarity/dissimilarity among SOM profiles.
}
