% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sommixplot.R
\name{sommix_star}
\alias{sommix_star}
\title{Plots SOM profiles using radial segment diagrams}
\usage{
sommix_star(
  som_obj,
  varnames = NULL,
  colormod = NULL,
  nodelab = TRUE,
  labtype = "IDs",
  labsize = 1,
  addFreq = FALSE,
  freqtype = "frq",
  freqsize = 1,
  legsymsize = 2,
  leglabsize = 1,
  legtxtlas = 2
)
}
\arguments{
\item{som_obj}{a SOM object}

\item{varnames}{specifies variable names}

\item{colormod}{can be used to customize plot color schemes. The default is terrain.colors().}

\item{nodelab}{whether or not to display reference labels aon each node}

\item{labtype}{determines type of node referencing to display. "IDs" and "XYs" are available}

\item{labsize}{sets size of node labels}

\item{addFreq}{whether or not to display frequency labels}

\item{freqtype}{determines type of frequency labels. Relative frequencies can be specified with "frq" or counts with "cnts"}

\item{freqsize}{frequency label cex value}

\item{legsymsize}{legend symbol cex value}

\item{leglabsize}{legend text cex value}

\item{legtxtlas}{legend text position rotation passed to las. Numeric in {0,1,2,3}; the style of axis label.}
}
\value{
a kohenen map where radial segment diagrams are used to illustrate profiles for each map node
}
\description{
Plots SOM profiles using radial segment diagrams
}
\details{
Radial segment diagrams provide a compact way to illustrate profiles derived from SOM. Here diagrams are presented along the SOM grid in order provide a compact visualization of SOM results.s the map.
}
