% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sommixsumm.R
\name{sommix_summ}
\alias{sommix_summ}
\title{SOM Summary Information}
\usage{
sommix_summ(som_obj)
}
\arguments{
\item{som_obj}{is a som object with class 'kohonen'}
}
\value{
'sommix'    a list of SOM results
\itemize{
 \item {som_grid} map specifics
 \item {som_coords} map reference coordinates system
 \item {som_codes} SOM profiles (a.k.a., cluster centers, classes, codebook vectors)
 \item {som_freq} SOM class frequencies summaries
 \item {som_class} a vector of SOM class assignments
 \item {trn_data} Training data merged with class assignments
 }
}
\description{
sommix_summ returns summary information for Kohonens self-organizing map algorithm
}
