% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compstats.R
\name{compstats}
\alias{compstats}
\title{Summary Statistics for observations assigned to each Self-Organizing Maps (SOM) node (i.e., map unit)}
\usage{
compstats(som_obj)
}
\arguments{
\item{som_obj}{an object returned from sommix}
}
\value{
a list containing summary statistics
}
\description{
Calculates summary statistics for data subsets assigned to each SOM node (i.e., map unit)
}
