% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sommixsumm.R
\name{sommix_summ}
\alias{sommix_summ}
\title{SOM Summary Information}
\usage{
sommix_summ(som_obj)
}
\arguments{
\item{som_obj}{is a som object with class 'kohonen'}
}
\value{
'sommix_summ'    a list of SOM results
\itemize{
 \item {PARAMETERS} map specifics
 \item {GRID}  map nodes and coordinates
 \item {CODES} node profiles (a.k.a., cluster centers, classes, codebook vectors)
 \item {FREQ} node assignment frequencies
 \item {CLASSIF} a vector identifying the best matching unit (BMU) on the SOM grid
 \item {DATA} input data merged with BMU assignments
 }
}
\description{
sommix_summ returns summary information for Kohonens self-organizing map algorithm
}
